add state updates plus new contract types
