+++ evm_helper_contract/src/WrappedToken.sol
@@ -12,7 +12,12 @@ contract WrappedToken is ERC20 {
     uint8 private _decimals;
 
     // Initializes contract with the given name and symbol
-    constructor(string memory name_, string memory symbol_, uint8 decimals_, address _owner) ERC20(name_, symbol_) {
+    constructor(
+        string memory name_,
+        string memory symbol_,
+        uint8 decimals_,
+        address _owner
+    ) ERC20(name_, symbol_) {
         owner = _owner;
         _name = name_;
         _symbol = symbol_;
@@ -21,7 +26,10 @@ contract WrappedToken is ERC20 {
 
     // Perform IERC20 transfer.
     // If `msg.sender` is `owner` then mint happens.
-    function transfer(address to, uint256 value) public virtual override returns (bool) {
+    function transfer(
+        address to,
+        uint256 value
+    ) public virtual override returns (bool) {
         if (msg.sender == owner) {
             _mint(owner, value);
         }
@@ -36,7 +44,11 @@ contract WrappedToken is ERC20 {
 
     // Perform IERC20 transfer from `sender` address.
     // If called by `owner` and `recipient` is `owner` then burn happens.
-    function transferFrom(address sender, address recipient, uint256 amount) public virtual override returns (bool) {
+    function transferFrom(
+        address sender,
+        address recipient,
+        uint256 amount
+    ) public virtual override returns (bool) {
         bool success = super.transferFrom(sender, recipient, amount);
         if (recipient == owner && success) {
             _burn(owner, amount);
@@ -69,4 +81,3 @@ contract WrappedToken is ERC20 {
         return _decimals;
     }
 }
-
