+++ minter/src/state/event.rs
@@ -2,7 +2,7 @@ use candid::Principal;
 use minicbor::{Decode, Encode};
 
 use crate::{
-    deposit_logs::{EventSource, ReceivedDepositEvent, ReceivedErc20Event, ReceivedNativeEvent},
+    contract_logs::{EventSource, ReceivedDepositEvent, ReceivedErc20Event, ReceivedNativeEvent},
     erc20::ERC20Token,
     eth_types::Address,
     lifecycle::{InitArg, UpgradeArg},
