+++ minter/src/tests/minter_flow_tets.rs
@@ -15,7 +15,7 @@ use crate::{
             five_ticks, icp_principal, lsm_principal, native_ledger_principal, update_call,
         },
     },
-    PROCESS_TOKENS_RETRIEVE_TRANSACTIONS_INTERVAL, SCRAPING_DEPOSIT_LOGS_INTERVAL,
+    PROCESS_TOKENS_RETRIEVE_TRANSACTIONS_INTERVAL, SCRAPING_contract_logs_INTERVAL,
 };
 
 use icrc_ledger_types::icrc1::account::Account;
@@ -718,7 +718,7 @@ fn should_deposit_and_withdrawal_erc20() {
         };
 
     pic.advance_time(
-        SCRAPING_DEPOSIT_LOGS_INTERVAL
+        SCRAPING_contract_logs_INTERVAL
             .checked_sub(Duration::from_secs(1 * 60))
             .unwrap(),
     );
